% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkmeans.R
\name{kkmeansTrain}
\alias{kkmeansTrain}
\title{Kernel k-means}
\usage{
kkmeansTrain(K, parameters)
}
\arguments{
\item{K}{kernel matrix}

\item{parameters}{list of parameters}
}
\description{
Perform the training step of kernel k-means
}
\references{
Gönen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).
}
\author{
Mehmet Gonen
}
