% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klic-package.R
\docType{package}
\name{klic-package}
\alias{klic}
\alias{klic-package}
\title{Kernel Learning Integrative Clustering}
\description{
Kernel Learning Integrative Clustering (KLIC) is an algorithm that allows to
combine multiple kernels, each representing a different measure of the
similarity between a set of observations. The contribution of each kernel on
the final clustering is weighted according to the amount of information
carried by it. As well as providing the functions required to perform the
kernel-based clustering, this package also allows the user to simply give the
data as input: the kernels are then built using consensus clustering.
Different strategies to choose the best number of clusters are also
available. For further details please see Cabassi and Kirk (2019)
<arXiv:1904.07701>.
}
\references{
Cabassi, A. and Kirk, P. D. W. (2019). Multiple kernel learning
for integrative consensus clustering of genomic datasets. arXiv preprint.
arXiv:1904.07701.

Gonen, M. and Margolin, A.A. (2014). Localized data fusion for
kernel k-means clustering with application to cancer biology. In Advances in
Neural Information Processing Systems (pp. 1305-1313).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/acabassi/klic}
  \item Report bugs at \url{http://github.com/acabassi/klic/issues}
}

}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk},
Paul D. W. Kirk,
Mehmet Gonen
}
