% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klic.R
\name{klic}
\alias{klic}
\title{Kernel learning integrative clustering}
\usage{
klic(data, M, individualK = NULL, individualMaxK = 6, globalK = NULL,
  globalMaxK = 6, B = 1000, C = 100, savePlots = FALSE,
  fileName = "klic")
}
\arguments{
\item{data}{List of M datasets, each of size N X P_m, m = 1, ..., M.}

\item{M}{number of data.}

\item{individualK}{Vector containing the number of clusters in each dataset. Default is NULL. If the number of clusters is not provided, then all the possible values between 2 and individualMaxK are considered and the best value is chosen for each dataset by maximising the silhouette.}

\item{individualMaxK}{Maximum number of clusters considered for the individual data. Default is 6.}

\item{globalK}{Number of global clusters. Default is NULL. If the number of clusters is not provided, then all the possible values between 2 and globalMaxK are considered and the best value is chosen by maximising the silhouette.}

\item{globalMaxK}{Maximum number of clusters considered for the final clustering. Default is 6.}

\item{B}{Number of iterations for consensus clustering. Default is 1000.}

\item{C}{Maximum number of iterations for localised kernel k-means. Default is 100.}

\item{savePlots}{If TRUE, a plot of the silhouette is saved in the working folder. Default is FALSE.}

\item{fileName}{If savePlots is TRUE, this is the name of the png file.}
}
\description{
Perform kernel learning integrative clustering
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
