% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmkkmeans.R
\name{lmkkmeans}
\alias{lmkkmeans}
\title{Localised multiple kernel k-means}
\usage{
lmkkmeans(Km, parameters, verbose = FALSE)
}
\arguments{
\item{Km}{kernel matrices}

\item{parameters}{list of parameters}

\item{verbose}{Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.}
}
\description{
Perform the training step of the localised multiple kernel k-means
}
\examples{
# Intialise 300 x 300 x 3 array containing M kernel matrices
# representing three different types of similarities
# between 300 data points
km <- array(NA, c(300, 300, 3))
# Load kernel matrices
km[,,1] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix1.csv", package = "klic"), row.names = 1))
km[,,2] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix2.csv", package = "klic"), row.names = 1))
km[,,3] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix3.csv", package = "klic"), row.names = 1))
#initalize the parameters of the algorithm
parameters <- list()
#set the number of clusters
parameters$cluster_count <- 2
# set the number of iterations
parameters$iteration_count <- 10
#perform training
state <- lmkkmeans(km, parameters)
#display the clustering
print(state$clustering)
#display the kernel weights
print(state$Theta)
}
\references{
Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).
}
\author{
Mehmet Gonen
}
