% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmkkmeans.R
\name{lmkkmeansTrain}
\alias{lmkkmeansTrain}
\title{Localised multiple kernel k-means}
\usage{
lmkkmeansTrain(Km, parameters, verbose = FALSE)
}
\arguments{
\item{Km}{kernel matrices}

\item{parameters}{list of parameters}

\item{verbose}{Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.}
}
\description{
Perform the training step of the localised multiple kernel k-means
}
\references{
Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).
}
\author{
Mehmet Gonen
}
