% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{maximiseSilhouette}
\alias{maximiseSilhouette}
\title{Choose K that maximises the silhouette from a kernel matrix and a list of clusterings}
\usage{
maximiseSilhouette(KM, clLabels, maxK, savePNG = FALSE,
  fileName = "silhouette")
}
\arguments{
\item{maxK}{Maximum number of clusters considered.}

\item{savePNG}{If TRUE, a plot of the silhouette is saved in the working folder. Defaults to FALSE.}

\item{fileName}{If savePNG is TRUE, this is the name of the png file.}

\item{kernelMatrix}{N X N X (maxK-1) array of kernel matrices.}

\item{clustering}{(maxK-1) X N matrix containing the clusterings obtained for different values of K}
}
\description{
Choose K that maximises the silhouette from a kernel matrix and a list of clusterings
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
