% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot silhouette}
\usage{
plotSilhouette(sil, fileName)
}
\arguments{
\item{sil}{vector of the average silhouette for K from 2 to some value maxK}

\item{fileName}{name of the png file}
}
\description{
Plot average silhouette.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
