% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-similarity-matrix.R
\name{plotSimilarityMatrix2}
\alias{plotSimilarityMatrix2}
\title{Plot similarity matrix with ggplot2}
\usage{
plotSimilarityMatrix2(CM, smTitle = NULL)
}
\arguments{
\item{smTitle}{plot title}

\item{SM}{similarity matrix}
}
\value{
ggplot2 plot
}
\description{
Plot similarity matrix using ggplot2.
}
