% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum-shift.R
\name{spectrumShift}
\alias{spectrumShift}
\title{Spectrum shift}
\usage{
spectrumShift(kernelMatrix, coeff = 1.2, verbose = FALSE)
}
\arguments{
\item{kernelMatrix}{symmetric matrix}
}
\description{
Make a symmetric matrix positive semi-definite
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
