% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophenetic-correlation.R
\name{copheneticCorrelation}
\alias{copheneticCorrelation}
\title{Cophenetic correlation coefficient}
\usage{
copheneticCorrelation(kernelMatrix)
}
\arguments{
\item{kernelMatrix}{kernel matrix.}
}
\description{
Compute the cophenetic correlation coefficient of a kernel matrix.
}
\references{
Sokal, R.R. and Rohlf, F.J., 1962. The comparison of dendrograms by objective methods. Taxon, 11(2), pp.33-40.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
