% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophenetic-correlation.R
\name{copheneticCorrelation}
\alias{copheneticCorrelation}
\title{Cophenetic correlation coefficient}
\usage{
copheneticCorrelation(kernelMatrix)
}
\arguments{
\item{kernelMatrix}{kernel matrix.}
}
\description{
Compute the cophenetic correlation coefficient of a kernel matrix, which is
a measure of howfaithfully hierarchical clustering would preserve the
pairwise distances between the original data points.
}
\examples{
# Load kernel matrix
km <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix1.csv", package = "klic"), row.names = 1))
# Compute cophenetic correlation
ccc <- copheneticCorrelation(km)
print(ccc)
}
\references{
Sokal, R.R. and Rohlf, F.J., 1962. The comparison of dendrograms by objective methods. Taxon, 11(2), pp.33-40.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
