% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkmeans.R
\name{kkmeans}
\alias{kkmeans}
\title{Kernel k-means}
\usage{
kkmeans(K, parameters)
}
\arguments{
\item{K}{Kernel matrix.}

\item{parameters}{A list containing the number of clusters
\code{number_count}.}
}
\value{
This function returns a list containing:
\item{clustering}{the cluster labels for each element (i.e. row/column) of
the kernel matrix.}
\item{objective}{the value of the objective function for the given
clustering.}
\item{parameters}{same parameters as in the input.}
}
\description{
Perform the training step of kernel k-means.
}
\references{
Gonen, M. and Margolin, A.A., 2014. Localized data fusion for
kernel k-means clustering with application to cancer biology. In Advances in
Neural Information Processing Systems (pp. 1305-1313).
}
\author{
Mehmet Gonen
}
