% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmkkmeans_missingData.R
\name{lmkkmeans_missingData}
\alias{lmkkmeans_missingData}
\title{Localised multiple kernel k-means}
\usage{
lmkkmeans_missingData(Km, parameters, missing = NULL, verbose = FALSE)
}
\arguments{
\item{Km}{Array of size N X N X M containing M kernel matrices.}

\item{parameters}{List of parameters, containing `iteration_count`, `cluster_count`.}

\item{missing}{Matrix of size N X M containing missingness indicators.}

\item{verbose}{Boolean flag. If TRUE, at each iteration the iteration number is printed.
Defaults to FALSE.}
}
\description{
Perform the training step of the localised multiple kernel k-means
}
\references{
Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means
clustering with application to cancer biology. In Advances in Neural Information Processing
Systems (pp. 1305-1313).
}
\author{
Mehmet Gonen, Alessandra Cabassi
}
