% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-similarity-matrix.R
\name{plotSimilarityMatrix}
\alias{plotSimilarityMatrix}
\title{Plot similarity matrix}
\usage{
plotSimilarityMatrix(X, y = NULL, clusLabels = NULL, colX = NULL,
  colY = NULL, myLegend = NULL, file_name = "myheatmap.png",
  savePNG = FALSE, saveTIKZ = FALSE, saveEPS = FALSE,
  semiSupervised = FALSE, scale = "none", labRow = NA, labCol = NA,
  dendro = "none")
}
\arguments{
\item{X}{matrix}

\item{y}{response}

\item{clusLabels}{cluster labels}

\item{colX}{colours for the matrix}

\item{colY}{colours for the response}

\item{myLegend}{vector of strings with the names of the variables}

\item{file_name}{name of the file}

\item{savePNG}{boolean flag. if true, saves the plot as a png file}

\item{saveTIKZ}{boolean flag. if true saves the plot as a tikz file}

\item{semiSupervised}{boolean flag. if true, the response is plotted next to the matrix.}

\item{scale}{parameter scale of heatmap.2. can be "none" or "columns"}

\item{labRow}{vector of row labels. default NA.}

\item{labCol}{vector of column labels. default NA.}
}
\description{
Plot similarity matrix. It is possible to plot a side vector that corresponds to a
response variable Y and to order the rows (and columns) according to some clustering
structure specified by the variable clusLabels.
}
