% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-similarity-matrix.R
\name{plotSimilarityMatrix}
\alias{plotSimilarityMatrix}
\title{Plot similarity matrix}
\usage{
plotSimilarityMatrix(X, y = NULL, clusLabels = NULL, colX = NULL,
  colY = NULL, myLegend = NULL, file_name = "myheatmap.png",
  savePNG = FALSE, saveTIKZ = FALSE, saveEPS = FALSE,
  semiSupervised = FALSE, scale = "none", labRow = NA, labCol = NA,
  dendro = "none")
}
\arguments{
\item{X}{Similarity matrix.}

\item{y}{Response variable.}

\item{clusLabels}{Cluster labels.}

\item{colX}{Colours for the matrix.}

\item{colY}{Colours for the response.}

\item{myLegend}{Vector of strings with the names of the variables.}

\item{file_name}{Name of the file.}

\item{savePNG}{Boolean flag: if TRUE, the plot is saved as a png file.}

\item{saveTIKZ}{Boolean flag: if TRUE, the plot is saved as a tikz file.}

\item{saveEPS}{Boolean flag: if TRUE, the plot is saved as an eps file.}

\item{semiSupervised}{Boolean flag: if TRUE, the response is plotted next to the matrix.}

\item{scale}{Used as input for the parameter "scale" of the gplot::heatmap.2() function.
Can be either "none" or "columns".}

\item{labRow}{Vector of row labels, default is NA.}

\item{labCol}{Vector of column labels, default is NA.}

\item{dendro}{If 'both', plot dendrogram on rows and columns; if 'none', no dendrograms are shown.
Default is 'none'.}
}
\description{
Plot similarity matrix. It is possible to plot a side vector that corresponds to a
response variable Y and to order the rows (and columns) according to some clustering
structure specified by the variable clusLabels.
}
\examples{
# Load one dataset with 300 observations, 2 variables, 6 clusters
data <- as.matrix(read.csv(system.file("extdata", "dataset1.csv",
package = "klic"), row.names = 1))
# Compute consensus clustering with K=6 clusters
cm <- coca::consensusCluster(data, 6)
# Plot consensus (similarity) matrix
plotSimilarityMatrix(cm)
}
