% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum-shift.R
\name{spectrumShift}
\alias{spectrumShift}
\title{Spectrum shift}
\usage{
spectrumShift(kernelMatrix, coeff = 1.2, verbose = FALSE)
}
\arguments{
\item{kernelMatrix}{symmetric matrix}

\item{coeff}{Coefficient by which the minimum eigenvalue is multiplied when shifting the eigenvalues, in order to avoid numeric problems. Default is 1.2.}

\item{verbose}{Boolean flag: if TRUE, information about the shift is printed to screen. Default is FALSE.}
}
\description{
Make a symmetric matrix positive semi-definite
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
